Regroup access control helpers in BaseCompanyController
