+++ app/controllers/CompanyAccessController.scala
@@ -11,38 +11,11 @@ import utils.silhouette.auth.AuthEnv
 
 @Singleton
 class CompanyAccessController @Inject()(
-                                companyRepository: CompanyRepository,
-                                companyAccessRepository: CompanyAccessRepository,
+                                val companyRepository: CompanyRepository,
+                                val companyAccessRepository: CompanyAccessRepository,
                                 val silhouette: Silhouette[AuthEnv]
                               )(implicit ec: ExecutionContext)
- extends BaseController {
-
-  // FIXME: Still experimental, will move
-  import com.mohiva.play.silhouette.api.actions.SecuredRequest
-  import utils.silhouette.auth.AuthEnv
-  import play.api.mvc._
-
-  type SecuredRequestWrapper[A] = SecuredRequest[AuthEnv, A]
-  class CompanyRequest[A](val company: Company, val accessLevel: AccessLevel, request: SecuredRequestWrapper[A]) extends WrappedRequest[A](request) {
-    def identity = request.identity
-  }
-  def withCompany[A](siret: String, authorizedLevels: Seq[AccessLevel])(implicit ec: ExecutionContext) = (SecuredAction andThen new ActionRefiner[SecuredRequestWrapper, CompanyRequest] {
-    def executionContext = ec
-    def refine[A](request: SecuredRequestWrapper[A]) = {
-      for {
-        company       <- companyRepository.findBySiret(siret)
-        accessLevel   <- company.map(
-                                    c => companyAccessRepository.getUserLevel(c.id, request.identity).map(Some(_)))
-                                .getOrElse(Future(None))
-      } yield {
-        company
-          .flatMap(c => accessLevel.map((c, _)))
-          .filter{case (_, l) => authorizedLevels.contains(l)}
-          .map{case (c, l) => Right(new CompanyRequest[A](c, l, request))}
-          .getOrElse(Left(NotFound))
-      }
-    }
-  })
+ extends BaseCompanyController {
 
   def listAccesses(siret: String) = withCompany(siret, List(AccessLevel.ADMIN)).async { implicit request =>
     for {
